Improve the release notes generator (#97)
This PR adds the ability to harvest community contributions automatically. To achieve that we need to provide the start commit. We can start inferring that commit from the tags if we wish. That change would be really simple, but I decided not to make too many changes for now.
Remove some unnecessary options. We've made parts of this tool generic in the past, but let's be real, it's super tailored to our repo and label management.
Added "TODO" for everything in the release notes that needs updated after you run the automatic generation.
Also added a bunch of things that can be inferred automatic, such as VS year, VS versions etc. This requires the occasional update of the tool, but better have to update the tool once every 2 years, then do manual like 8-9 over a few years.