HTML encode issue titles in release notes tool
If someone reports an issue on a specific method, and that method is async
the compiler generates a name like `TypeName+<Method>`, but that looks
like an invalid HTML tag that will get hidden when rendered. Therefore, HTML
encode it, so it gets displayed correctly.

Also update the project to .NET 8, update packages and fix a few compiler warnings